requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.13.4

# Tools
django-extensions==3.1.0

# Testing
mypy==0.790
pytest==6.2.1
pytest-sugar==0.9.4
pytest-django==4.1.0
factory-boy==3.2.0

# Code quality
flake8==3.8.4